src/requirements.txt
pinecone-client
openai
langchain
tqdm
unstructured
pdf2image
tiktoken
flask
slack_sdk
